Conflicts:
	Repository.mk
	RepositoryFixes.mk
	configure
	configure.in
	set_soenv.in
	solenv/config/sdev300.ini
	solenv/gbuild/ComponentTarget.mk
	solenv/gbuild/CppunitTest.mk
	solenv/gbuild/CustomTarget.mk
	solenv/gbuild/Deliver.mk
	solenv/gbuild/Executable.mk
	solenv/gbuild/JavaClassSet.mk
	solenv/gbuild/Library.mk
	solenv/gbuild/LinkTarget.mk
	solenv/gbuild/StaticLibrary.mk
	solenv/gbuild/TargetLocations.mk
	solenv/gbuild/gbuild.mk
	solenv/gbuild/platform/linux.mk
	solenv/gbuild/platform/macosx.mk
	solenv/gbuild/platform/solaris.mk
	solenv/gbuild/platform/windows.mk
	solenv/gbuild/platform/winmingw.mk
	solenv/gbuild/templates/Executable.mk
	solenv/gbuild/templates/Library.mk
	solenv/inc/libs.mk
	solenv/inc/unxobsd.mk

+++ soltools/cpp/_cpp.c
@@ -1,3 +1,4 @@
+/* -*- Mode: C++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*- */
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -298,7 +299,7 @@ void
 }
 
 void *
-    domalloc(int size)
+    domalloc(size_t size)
 {
     void *p = malloc(size);
 
@@ -381,3 +382,5 @@ void
         nerrs = 1;
     fflush(stderr);
 }
+
+/* vim:set shiftwidth=4 softtabstop=4 expandtab: */
